+++ ruby/lib/google/protobuf_ffi.rb
@@ -28,44 +28,54 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-require 'ffi'
-require 'google/protobuf/ffi/internal/type_safety'
-require 'google/protobuf/ffi/internal/arena'
-require 'google/protobuf/ffi/internal/convert'
-require 'google/protobuf/ffi/descriptor'
-require 'google/protobuf/ffi/enum_descriptor'
-require 'google/protobuf/ffi/field_descriptor'
-require 'google/protobuf/ffi/oneof_descriptor'
-require 'google/protobuf/ffi/ffi'
-require 'google/protobuf/ffi/descriptor_pool'
-require 'google/protobuf/ffi/file_descriptor'
-require 'google/protobuf/ffi/map'
-require 'google/protobuf/ffi/object_cache'
-require 'google/protobuf/ffi/repeated_field'
-require 'google/protobuf/descriptor_dsl'
+# require mixins before we hook them into the java & c code
+require 'google/protobuf/message_exts'
 
+# We define these before requiring the platform-specific modules.
+# That way the module init can grab references to these.
 module Google
   module Protobuf
-    def self.deep_copy(object)
-      case object
-      when RepeatedField
-        RepeatedField.send(:deep_copy, object)
-      when Google::Protobuf::Map
-        Google::Protobuf::Map.deep_copy(object)
-      when Google::Protobuf::MessageExts
-        object.class.send(:deep_copy, object.instance_variable_get(:@msg))
-      else
-        raise NotImplementedError
-      end
+    class Error < StandardError; end
+    class ParseError < Error; end
+    class TypeError < ::TypeError; end
+
+    begin
+      require 'ffi'
+      PREFER_FFI = true
+    rescue LoadError
+      PREFER_FFI = false
     end
 
-    def self.discard_unknown(message)
-      raise FrozenError if message.frozen?
-      raise ArgumentError.new "Expected message, got #{message.class} instead." if message.instance_variable_get(:@msg).nil?
-      unless Google::Protobuf::FFI.message_discard_unknown(message.instance_variable_get(:@msg), message.class.descriptor, 128)
-        raise RuntimeError.new "Messages nested too deeply."
+    def self.encode(msg, options = {})
+      msg.to_proto(options)
+    end
+
+    def self.encode_json(msg, options = {})
+      msg.to_json(options)
+    end
+
+    def self.decode(klass, proto, options = {})
+      klass.decode(proto, options)
+    end
+
+    def self.decode_json(klass, json, options = {})
+      klass.decode_json(json, options)
+    end
+
+    IMPLEMENTATION = if PREFER_FFI
+      begin
+        require 'google/protobuf_ffi'
+        :FFI
+      rescue LoadError
+        warn "Caught exception `#{$!.message}` while loading FFI implementation of google/protobuf."
+        raise $!
+        warn "Falling back to native implementation."
+        require 'google/protobuf_native'
+        :NATIVE
       end
-      nil
+    else
+      require 'google/protobuf_native'
+      :NATIVE
     end
   end
 end
